# Src/Photonne.Server.Api/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Photonne API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
  "paths": {
...
    "/api/assets/{assetId}/same-day": {
      "get": {
        "tags": [
          "Assets"
        ],
        "description": "Returns other visible assets captured on the same calendar day as the given asset",
        "operationId": "GetSameDayAssets",
        "parameters": [
          {
            "name": "assetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SameDayAssetsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/api/assets/{assetId}": {
      "get": {
        "tags": [
          "Assets"
        ],
        "summary": "Get asset details",
        "description": "Returns detailed information about an asset including EXIF data, thumbnails, tags, and folder information.",
        "operationId": "GetAssetDetail",
        "parameters": [
          {
            "name": "assetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "required": [
          "error",
          "code"
        ],
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      },
...
      "PersonAssetDto": {
        "required": [
          "id",
          "fileName",
          "type",
          "fileCreatedAt",
          "hasThumbnails",
          "dominantColor"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fileName": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "fileCreatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "hasThumbnails": {
            "type": "boolean"
          },
          "dominantColor": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SameDayAssetsResponse": {
        "required": [
          "total",
          "items"
        ],
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PersonAssetDto"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Access token from /api/auth/login or /api/auth/refresh",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}